refactor: generialise FAQItem & FAQAccordionProps, keep notion faqs in notion compare page
