+++ app/pricing/components/faq-accordion.tsx
@@ -3,18 +3,8 @@
 import React, { useState, useEffect } from 'react';
 import { motion } from 'framer-motion';
 import { usePricingState } from './pricing-state-context';
+import { FAQAccordionProps } from '@/lib/faq';
 
-interface FAQItem {
-  id: string;
-  question: string;
-  answer: string | React.ReactNode;
-}
-
-interface FAQAccordionProps {
-  items: FAQItem[];
-}
-
-// 展开时的减号图标
 function MinusIcon({ isHovered }: { isHovered: boolean }) {
   return (
     <motion.svg
@@ -104,19 +94,18 @@ export function FAQAccordion({ items }: FAQAccordionProps) {
         return (
           <motion.div
             key={item.id}
-            className={`flex cursor-pointer flex-col justify-center rounded-[15px] border-2 bg-white px-6 py-6 sm:px-8 sm:py-8 md:px-[40px] md:py-[40px] select-none touch-manipulation ${
-              isExpanded ? 'gap-3 sm:gap-4 md:gap-[15px]' : isAnimating ? 'gap-3' : 'gap-0'
-            }`}
+            className={`flex cursor-pointer flex-col justify-center rounded-[15px] border-2 bg-white px-6 py-6 sm:px-8 sm:py-8 md:px-[40px] md:py-[40px] select-none touch-manipulation ${isExpanded ? 'gap-3 sm:gap-4 md:gap-[15px]' : isAnimating ? 'gap-3' : 'gap-0'
+              }`}
             style={{
               WebkitTapHighlightColor: 'transparent',
               ...(isHovered
                 ? {
-                    borderColor: 'transparent',
-                    background: `linear-gradient(white, white) padding-box, ${getGradient()} border-box`,
-                  }
+                  borderColor: 'transparent',
+                  background: `linear-gradient(white, white) padding-box, ${getGradient()} border-box`,
+                }
                 : {
-                    borderColor: 'rgba(213, 215, 222, 0.40)',
-                  })
+                  borderColor: 'rgba(213, 215, 222, 0.40)',
+                })
             }}
             animate={{
               borderColor: isHovered ? 'transparent' : 'rgba(213, 215, 222, 0.40)',
