+++ lib/faq.ts
@@ -1,0 +1,9 @@
+export interface FAQItem {
+  id: string;
+  question: string;
+  answer: string | React.ReactNode;
+}
+
+export interface FAQAccordionProps {
+  items: FAQItem[];
+}
